+++ src/application/useCases/sendNotification.spec.ts
@@ -1,26 +1,20 @@
-import { Notification } from '../entities/notification';
+import { InMemoryNotificationsRepository } from '../../../test/repositories/inMemoryNotificationsRepository';
 import { SendNotification } from './sendNotification';
-
-const notifications: Notification[] = [];
-
-const notificationsRepository = {
-  async create(notification: Notification) {
-    notifications.push(notification);
-  },
-};
 
 describe('Send notification', () => {
   it('should be able to send a notification', async () => {
+    const notificationsRepository = new InMemoryNotificationsRepository();
     const sendNotification = new SendNotification(notificationsRepository);
 
-    await sendNotification.execute({
+    const { notification } = await sendNotification.execute({
       content: 'This is a notification',
       category: 'social',
       recepientId: 'example-recepient-id',
     });
 
-    console.log(notifications);
+    console.log(notificationsRepository.notifications);
 
-    expect(notifications).toHaveLength(1);
+    expect(notificationsRepository.notifications).toHaveLength(1);
+    expect(notificationsRepository.notifications[0]).toEqual(notification);
   });
 });