+++ changelog.md
@@ -6,6 +6,7 @@
 - Send notification use case
 - Tests for notification and content
 - Tests for sendNotification useCase
+- In-memory database for tests
 
 ### [0.0.2] - 2022-12-12